# src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: Winbou Prices API
  version: 1.0.0
  description: REST API to query the applicable price of a product on a given date, according to the rates defined in the PRICES table.
servers:
  - url: http://localhost:8080
    description: Local server

paths:
  /prices:
    get:
      summary: Retrieves the applicable price for a brand on a given date
      operationId: getApplicablePrice
      parameters:
        - in: query
          name: applicationDate
          required: true
          description: Application date (2020-06-14T10:00:00)
          schema:
            type: string
            format: date-time
        - in: query
          name: productId
          required: true
          description: Product identifier
          schema:
            type: integer
            format: int64
        - in: query
          name: brandId
          required: true
          description: Brand identifier
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Applicable price found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidParameterError'
        '404':
          description: No applicable price found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceNotFoundError'

components:
  schemas:
    PriceResponse:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          example: 35455
        brandId:
          type: integer
          format: int64
          example: 1
        priceList:
          type: integer
          example: 2
        startDate:
          type: string
          format: date-time
          example: 2020-06-14T15:00:00
        endDate:
          type: string
          format: date-time
          example: 2020-06-14T18:30:00
        price:
          type: number
          format: double
          example: 25.45
        currency:
          type: string
          example: EUR
    InvalidParameterError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Invalid parameter applicationDate must be in ISO-8601 format
        path:
          type: string
          example: /prices
    PriceNotFoundError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: No applicable price found for brand=1, product=35455
        path:
          type: string
          example: /prices